Moved items from settings.py to consts.py.
