+++ chat_bot/utils/consts.py
@@ -1,0 +1,14 @@
+COMMAND_STR: str = '17!'
+DATA_QUERY_L: str = '{{'
+DATA_QUERY_R: str = '}}'
+DATA_QUERY_MID: str = '|'
+QUOTE_PAIRS: dict[str, str] = {'"': '"', '“': '”'}
+
+FORMAT_NICKNAMES: dict[str, str] = {
+    'PremierDraft': 'Bo1',
+    'TradDraft': 'Bo3',
+    'QuickDraft': 'Quick',
+    'Sealed': 'Seal',
+    'TradSealed': 'Bo3Seal',
+    'DraftChallenge': 'Chal.',
+}
