+++ chat_bot/utils/settings.py
@@ -1,23 +1,9 @@
-COMMAND_STR: str = '17!'
-DATA_QUERY_L: str = '{{'
-DATA_QUERY_R: str = '}}'
-DATA_QUERY_MID: str = '|'
-DEFAULT_FORMAT: str = 'PremierDraft'
-START_DATE: str = '2019-01-01'
-QUOTE_PAIRS: dict[str, str] = {'"': '"', '“': '”'}
-
-FORMAT_NICKNAMES: dict[str, str] = {
-    'PremierDraft': 'Bo1',
-    'TradDraft': 'Bo3',
-    'QuickDraft': 'Quick',
-    'Sealed': 'Seal',
-    'TradSealed': 'Bo3Seal',
-    'DraftChallenge': 'Chal.',
-}
-
 UPDATING_SETS: list[str] = ['NEO']
 OLD_SETS: list[str] = []  # ['DBL', 'VOW', 'MID', 'AFR', 'STX', 'KHM']
 SETS: list[str] = ['NEO']  # UPDATING_SETS + OLD_SETS
+START_DATE: str = '2019-01-01'
+DEFAULT_FORMAT: str = 'PremierDraft'
+
 
 FORMATS: dict[str, list[str]] = {
     'PremierDraft': ['bo1', 'premier', 'premierdraft'],
